Installing React Router and creating new pages
- Installing React Router.
- Create the Dashbard page for the root of the Home path.
- Create a RootPage as the roo path of the application.
